+++ libs/full/async_distributed/include/hpx/async_distributed/applier/apply.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2007-2018 Hartmut Kaiser
+//  Copyright (c) 2007-2020 Hartmut Kaiser
 //  Copyright (c)      2011 Bryce Lelbach
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -13,7 +13,7 @@
 #include <hpx/actions_base/actions_base_support.hpp>
 #include <hpx/actions_base/traits/extract_action.hpp>
 #include <hpx/assert.hpp>
-#include <hpx/async_distributed/applier/detail/apply_implementations.hpp>
+#include <hpx/async_distributed/applier/detail/apply_implementations_fwd.hpp>
 #include <hpx/async_local/apply.hpp>
 #include <hpx/components_base/traits/component_type_is_compatible.hpp>
 #include <hpx/functional/traits/is_action.hpp>
@@ -23,7 +23,8 @@
 #include <hpx/runtime/agas/interface.hpp>
 #include <hpx/runtime/components/client_base.hpp>
 #include <hpx/runtime/parcelset/detail/parcel_await.hpp>
-#include <hpx/runtime/parcelset/put_parcel.hpp>
+#include <hpx/runtime/parcelset/parcel.hpp>
+#include <hpx/runtime/parcelset/put_parcel_fwd.hpp>
 #include <hpx/runtime/parcelset_fwd.hpp>
 #include <hpx/runtime_local/get_locality_id.hpp>
 #include <hpx/traits/is_continuation.hpp>
@@ -679,3 +680,8 @@ namespace hpx {
             gid, actions::action_priority<Derived>(), std::forward<Ts>(vs)...);
     }
 }    // namespace hpx
+
+// these files are intentionally #included last as it refers to functions
+// defined here
+#include <hpx/async_distributed/applier/detail/apply_implementations.hpp>
+#include <hpx/runtime/parcelset/put_parcel.hpp>
